Remove warning for missing OPENAI_API_KEY
Removed check for missing OPENAI_API_KEY environment variable from error handling block in the app.py file. As the key warning was producing noise in the logs and was wasn't necessary for the app's runtime. It was also distracting as all required environment variables are now ensured to be set during the CDK/Docker setup process. No impact on weather icon and OpenAI prompt setup.